修复字幕问题
