+++ src/fronted/components/srt-cops/atoms/word-pop.tsx
