+++ src/fronted/components/srt-cops/translatable-line-podcast.tsx
@@ -3,7 +3,7 @@ import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/fron
 import { Button } from '@/fronted/components/ui/button';
 import { cn } from '@/fronted/lib/utils';
 import { Sentence } from '@/common/types/SentenceC';
-import TranslatableLineCore from '@/fronted/components/srt-cops/atoms/TranslatableLineCore';
+import TranslatableLineCore from '@/fronted/components/srt-cops/atoms/translatable-line-core';
 
 interface TranslatableLinePodcastParam {
     sentence: Sentence;
